+++ tests/Config/SyncConfigTest.php
@@ -10,6 +10,7 @@ use Grixu\Synchronizer\Tests\Helpers\FakeLoader;
 use Grixu\Synchronizer\Tests\Helpers\FakeParser;
 use Grixu\Synchronizer\Tests\Helpers\FakeSyncConfig;
 use Grixu\Synchronizer\Tests\Helpers\TestCase;
+use Illuminate\Queue\SerializableClosure;
 use Illuminate\Support\Collection;
 use Throwable;
 
@@ -18,20 +19,29 @@ class SyncConfigTest extends TestCase
     /** @test */
     public function it_could_take_every_argument()
     {
-        $obj = new SyncConfig(
+        $obj = $this->createObj();
+
+        $this->assertEquals(SyncConfig::class, $obj::class);
+    }
+
+    protected function createObj(): SyncConfig
+    {
+        return new SyncConfig(
             loaderClass: FakeLoader::class,
             parserClass: FakeParser::class,
             localModel: Language::class,
             foreignKey: 'xlId',
             jobsConfig: config('synchronizer.jobs.default'),
             idsToSync: null,
-            syncClosure: function (Collection $dtoCollection, SyncConfig $config) {
-        },
-            errorHandler: function (Throwable $e) {
-        }
+            syncClosure: new SerializableClosure(
+                             function (Collection $dtoCollection, SyncConfig $config) {
+                             }
+                         ),
+            errorHandler: new SerializableClosure(
+                             function (Throwable $e) {
+                             }
+                         )
         );
-
-        $this->assertEquals(SyncConfig::class, $obj::class);
     }
 
     /** @test */
@@ -45,10 +55,14 @@ class SyncConfigTest extends TestCase
                 foreignKey: 'xlId',
                 jobsConfig: config('synchronizer.jobs.default'),
                 idsToSync: null,
-                syncClosure: function (Collection $dtoCollection, SyncConfig $config) {
-            },
-                errorHandler: function (Throwable $e) {
-            }
+                syncClosure: new SerializableClosure(
+                                 function (Collection $dtoCollection, SyncConfig $config) {
+                                 }
+                             ),
+                errorHandler: new SerializableClosure(
+                                 function (Throwable $e) {
+                                 }
+                             )
             );
 
             $this->assertTrue(false);
@@ -76,10 +90,32 @@ class SyncConfigTest extends TestCase
         $config = FakeSyncConfig::make();
 
         try {
-            $config->setCurrentJob(count(config('synchronizer.jobs.default'))+1);
+            $config->setCurrentJob(count(config('synchronizer.jobs.default')) + 1);
             $this->assertTrue(false);
         } catch (Exception) {
             $this->assertTrue(true);
         }
+    }
+
+    /** @test */
+    public function it_convert_closure_to_serializable_closure_on_sync_handler_setter()
+    {
+        $config = $this->createObj();
+
+        $config->setSyncClosure(function () {});
+
+        $this->assertNotEmpty($config->getSyncClosure());
+        $this->assertEquals(SerializableClosure::class, $config->getSyncClosure()::class);
+    }
+
+    /** @test */
+    public function it_convert_closure_to_serializable_closure_on_error_handler_setter()
+    {
+        $config = $this->createObj();
+
+        $config->setErrorHandler(function () {});
+
+        $this->assertNotEmpty($config->getErrorHandler());
+        $this->assertEquals(SerializableClosure::class, $config->getErrorHandler()::class);
     }
 }